+++ AnorocMobileApp/AnorocMobileApp/Models/Notification/NotificationModel.cs
@@ -1,4 +1,5 @@
-﻿using System.ComponentModel;
+﻿using System;
+using System.ComponentModel;
 using System.Runtime.CompilerServices;
 using System.Runtime.Serialization;
 using Xamarin.Forms.Internals;
@@ -10,6 +11,7 @@ namespace AnorocMobileApp.Models.Notification
     /// </summary>
     [Preserve(AllMembers = true)]
     [DataContract]
+    [Serializable]
     public class NotificationModel : INotifyPropertyChanged
     {
 
@@ -30,7 +32,7 @@ namespace AnorocMobileApp.Models.Notification
         /// <summary>
         /// Gets or sets the received time of an item.
         /// </summary>
-        [DataMember(Name = "receivedTime")]
+        [DataMember(Name = "receivedTime")]        
         public string ReceivedTime { get; set; }
 
         /// <summary>
